+++ src/config/index.ts
@@ -6,10 +6,12 @@ import { db, DbConfig } from './database'
 import { docs, DocsConfig } from './docs'
 import { log, LogConfig } from './log'
 import { mail, MailConfig } from './mail'
+import { oauth, OAuthConfig } from './oauth'
 import { search, SearchConfig } from './search'
 import { schedule, ScheduleConfig } from './schedule'
 import { session, SessionConfig } from './session'
 import { socket, SocketConfig } from './socket'
+import { totp, TotpConfig } from './totp'
 import { webhook, WebhookConfig } from './webhook'
 
 export const config: ConfigDictionary = {
@@ -20,10 +22,12 @@ export const config: ConfigDictionary = {
 	docs,
 	log,
 	mail,
+	oauth,
 	search,
 	session,
 	schedule,
 	socket,
+	totp,
 	webhook,
 }
 
@@ -35,9 +39,11 @@ interface ConfigDictionary {
 	docs: DocsConfig
 	log: LogConfig
 	mail: MailConfig
+	oauth: OAuthConfig
 	search: SearchConfig
 	session: SessionConfig
 	schedule: ScheduleConfig
 	socket: SocketConfig
+	totp: TotpConfig
 	webhook: WebhookConfig
 }